demo_agent_1/agent/requirements.txt
langgraph
langchain_community
langchain_groq
langchain_openai
pydantic
sqlparse
psycopg2-binary